feat: use std sort
